Remove old DC exporter
